requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
fastapi==0.114.0
greenlet==3.0.3
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.9.0
pydantic_core==2.23.2
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1

uvicorn~=0.30.6
PyJWT~=2.9.0